+++ Mimir.MongoDB/Bson/Serialization/SerializationRegistry.cs
@@ -1,5 +1,5 @@
 using Libplanet.Crypto;
-using Mimir.MongoDB.Bson.Serialization.Serializers;
+using Mimir.MongoDB.Bson.Serialization.Serializers.Libplanet;
 using MongoDB.Bson.Serialization;
 
 namespace Mimir.MongoDB.Bson.Serialization;
@@ -8,6 +8,7 @@ public static class SerializationRegistry
 {
     public static void Register()
     {
-        BsonSerializer.RegisterSerializer(typeof(Address), new LibplanetCryptoAddressBsonSerializer());
+        // Libplanet
+        BsonSerializer.RegisterSerializer(typeof(Address), new AddressSerializer());
     }
 }
