+++ Mimir.MongoDB/Bson/Serialization/Serializers/Libplanet/AddressSerializer.cs
@@ -4,9 +4,9 @@ using MongoDB.Bson;
 using MongoDB.Bson.Serialization;
 using MongoDB.Bson.Serialization.Serializers;
 
-namespace Mimir.MongoDB.Bson.Serialization.Serializers;
+namespace Mimir.MongoDB.Bson.Serialization.Serializers.Libplanet;
 
-public class LibplanetCryptoAddressBsonSerializer : SerializerBase<Address>
+public class AddressSerializer : StructSerializerBase<Address>
 {
     public override void Serialize(BsonSerializationContext context, BsonSerializationArgs args, Address value) =>
         context.Writer.WriteString(value.ToHex());
